testando transmissao entre NoCs

+++ peripheral/secNoC/secRouter.c
@@ -167,7 +167,7 @@ PPM_PACKETNET_CB(unsafeNoC) {
 
         incomingFlit.source = myID;
         incomingFlit.data = *(unsigned int *) data;
-        incomingFlit.data = htonl1((unsigned int)incomingFlit.data);
+        incomingFlit.data = htonl1(incomingFlit.data);
         bhmMessage("INFO", "SECROUTER", "-------------------------------------------------------------> flitTotal = %d received", incomingFlit.data);
         bufferPush(LOCAL);
